backend/requirements.txt
fastapi
uvicorn
pydantic-settings
pillow
tensorflow
numpy